backend/requirements_venv.txt
aiogrpc==1.8
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
bidict==0.23.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.0
fabric-sdk-py==0.8.1
fastapi==0.119.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
granian==2.5.5
greenlet==3.2.4
grpcio==1.75.1
h11==0.16.0
hkdf==0.0.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
lark-parser==0.7.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==25.0
platformdirs==4.5.0
protobuf==3.20.3
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pysha3==1.0b1
python-dotenv==1.1.1
python-engineio==4.12.3
python-multipart==0.0.20
python-socketio==5.14.2
PyYAML==6.0.3
redis==6.4.0
reflex==0.8.15
reflex-hosting-cli==0.1.57
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
Rx==3.2.0
sentry-sdk==2.42.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.48.0
tomli==2.3.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wrapt==1.17.3
wsproto==1.2.0
